[REFACTOR] 술 약속 관련 JWT 수정 #66

+++ src/main/java/umc/puppymode/service/DrinkingAppointmentService/DrinkingAppointmentCommendServiceImpl.java
@@ -53,43 +53,75 @@ public class DrinkingAppointmentCommendServiceImpl implements DrinkingAppointmen
 
     @Override
     @Transactional
-    public void deleteDrinkingAppointment(Long appointmentId) {
+    public void deleteDrinkingAppointment(Long appointmentId, Long userId) {
+
+        // User 엔티티 조회
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 사용자 ID 입니다."));
+
         DrinkingAppointment appointment = repository.findById(appointmentId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 약속을 찾을 수 없습니다."));
+
+        // 약속 소유자 검증
+        if (!appointment.getUser().getUserId().equals(user.getUserId())) {
+            throw new IllegalStateException("해당 약속을 삭제할 권한이 없습니다.");
+        }
 
         repository.delete(appointment);
     }
 
     @Override
     @Transactional
-    public void rescheduleDrinkingAppointment(Long appointmentId, DrinkingAppointmentRequestDTO.RescheduleAppointmentRequestDTO request) {
+    public void rescheduleDrinkingAppointment(Long appointmentId, DrinkingAppointmentRequestDTO.RescheduleAppointmentRequestDTO request, Long userId) {
+
+        // User 엔티티 조회
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 사용자 ID 입니다."));
 
         //약속 조회
         DrinkingAppointment appointment = repository.findById(appointmentId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 약속을 찾을 수 없습니다."));
+
+        // 약속 소유자 검증
+        if (!appointment.getUser().getUserId().equals(user.getUserId())) {
+            throw new IllegalStateException("해당 약속을 미룰 권한이 없습니다.");
+        }
 
         // 시간 유효성 검증
         if (request.getDateTime().isBefore(LocalDateTime.now())) {
             throw new IllegalArgumentException("현재 시간 이전으로 설정할 수 없습니다.");
         }
 
+        validateAppointmentStatusForRescheduling(appointment);
+
+        appointment.setDateTime(request.getDateTime());
+    }
+
+    private static void validateAppointmentStatusForRescheduling(DrinkingAppointment appointment) {
         // 술 약속 상태 검증
         if (appointment.getStatus() == AppointmentStatus.ONGOING){
             throw new IllegalArgumentException("이미 진행 중인 약속은 미룰 수 없습니다.");
         } else if (appointment.getStatus() == AppointmentStatus.COMPLETED) {
             throw new IllegalArgumentException("이미 완료 된 약속은 미를 수 없습니다.");
         }
-
-        appointment.setDateTime(request.getDateTime());
     }
 
     @Override
     @Transactional
-    public void completeDrinkingAppointment(Long appointmentId) {
+    public void completeDrinkingAppointment(Long appointmentId, Long userId) {
+
+        // User 엔티티 조회
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 사용자 ID 입니다."));
 
         // 약속 조회
         DrinkingAppointment appointment = repository.findById(appointmentId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 약속을 찾을 수 없습니다."));
+
+        // 약속 소유자 검증
+        if (!appointment.getUser().getUserId().equals(user.getUserId())) {
+            throw new IllegalStateException("음주를 종료할 권한이 없습니다.");
+        }
 
         // 해당 약속 상태가 ONGOING 상태 인지 확인
         AppointmentStatus status = appointment.getStatus();
